CP-54351 move ipc from /opt
